Issue #202: Reduced the file length of InputMutationTests15
# Conflicts:
#	config/checkstyle-resources-suppressions.xml
#	src/test/java/org/checkstyle/autofix/recipe/UseEnhancedSwitchTest.java
