+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests15/DiffMutationTests15.diff
@@ -1,6 +1,6 @@
 --- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests15/InputMutationTests15.java
 +++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests15/OutputMutationTests15.java
-@@ -7,127 +7,92 @@
+@@ -7,99 +7,70 @@
  */
  package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests15;
  
@@ -129,148 +129,6 @@
 -                break;
 -            }
 +            default -> System.out.println("c");
-         }
-     }
- 
-     void testContinueControlFlow(int x) {
-         for (int i = 0; i < 10; i++) {
-             int y;
--            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--                case 1:
-+            switch (x) {
-+                case 1 -> {
-                     if (i > 5) continue;
--                    y = 1;
--                    break;
--                default:
--                    y = 0;
--                    break;
-+                    y = 1;}
-+                default -> y = 0;
-             }
-         }
-     }
- 
-     void testReturnControlFlowInAssignment(int x) {
-         int y;
--        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--            case 1:
-+        switch (x) {
-+            case 1 -> {
-                 if (x > 100) return;
--                y = 1;
--                break;
--            default:
--                y = 0;
--                break;
-+                y = 1;}
-+            default -> y = 0;
-         }
-     }
- 
-@@ -135,81 +100,57 @@
-         outer:
-         for (int i = 0; i < 5; i++) {
-             int y;
--            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--                case 1:
-+            switch (x) {
-+                case 1 -> {
-                     if (x > 100) break outer;
--                    y = 1;
--                    break;
--                default:
--                    y = 0;
--                    break;
-+                    y = 1;}
-+                default -> y = 0;
-             }
-         }
-     }
- 
-     void testBlockIndent(int x) {
--        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--            case 1: {
-+        switch (x) {
-+            case 1 -> {
-                     int y = 1;
-                     System.out.println(y);
--                    break;
-                 }
--            default: {
-+            default -> {
-                     int y = 0;
-                     System.out.println(y);
--                    break;
-                 }
-         }
-     }
- 
-     void testAssignmentBlockYield(int x) {
--        int y;
--        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--            case 1: {
-+        int y = switch (x) {
-+            case 1 -> {
-                 int z = x + 1;
--                y = z;
--                break;
-+                yield z;
-             }
--            default:
--                y = 0;
--                break;
--        }
-+            default -> 0;
-+        };
-     }
- 
-     void testSingleAssignment(int x) {
--        int y;
--        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--            case 1:
--                y = 100;
--                break;
--            case 2:
--                y = 200;
--                break;
--            default:
--                y = 0;
--                break;
--        }
-+        int y = switch (x) {
-+            case 1 -> 100;
-+            case 2 -> 200;
-+            default -> 0;
-+        };
-     }
- 
-     int testCollectHandled(Color c) {
--        switch (c) { // violation 'Switch can be replaced with enhanced switch.'
--            case RED:
--            case GREEN:
--                return 1;
--            case BLUE:
--                return 2;
--            default:
--                return 0;
--        }
-+        return switch (c) {
-+            case RED, GREEN -> 1;
-+            case BLUE -> 2;
-+        };
-     }
- 
-     void testEmptyDefault(int x) {
--        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
--            case 1:
--                System.out.println("one");
--                break;
--            default:
--                break;
-+        switch (x) {
-+            case 1 -> System.out.println("one");
-+            default -> {}
          }
      }
  }
