+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests15/InputMutationTests15.java
@@ -102,114 +102,4 @@ public class InputMutationTests15 {
             }
         }
     }
-
-    void testContinueControlFlow(int x) {
-        for (int i = 0; i < 10; i++) {
-            int y;
-            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-                case 1:
-                    if (i > 5) continue;
-                    y = 1;
-                    break;
-                default:
-                    y = 0;
-                    break;
-            }
-        }
-    }
-
-    void testReturnControlFlowInAssignment(int x) {
-        int y;
-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-            case 1:
-                if (x > 100) return;
-                y = 1;
-                break;
-            default:
-                y = 0;
-                break;
-        }
-    }
-
-    void testLabeledBreakControlFlowInAssignment(int x) {
-        outer:
-        for (int i = 0; i < 5; i++) {
-            int y;
-            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-                case 1:
-                    if (x > 100) break outer;
-                    y = 1;
-                    break;
-                default:
-                    y = 0;
-                    break;
-            }
-        }
-    }
-
-    void testBlockIndent(int x) {
-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-            case 1: {
-                    int y = 1;
-                    System.out.println(y);
-                    break;
-                }
-            default: {
-                    int y = 0;
-                    System.out.println(y);
-                    break;
-                }
-        }
-    }
-
-    void testAssignmentBlockYield(int x) {
-        int y;
-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-            case 1: {
-                int z = x + 1;
-                y = z;
-                break;
-            }
-            default:
-                y = 0;
-                break;
-        }
-    }
-
-    void testSingleAssignment(int x) {
-        int y;
-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-            case 1:
-                y = 100;
-                break;
-            case 2:
-                y = 200;
-                break;
-            default:
-                y = 0;
-                break;
-        }
-    }
-
-    int testCollectHandled(Color c) {
-        switch (c) { // violation 'Switch can be replaced with enhanced switch.'
-            case RED:
-            case GREEN:
-                return 1;
-            case BLUE:
-                return 2;
-            default:
-                return 0;
-        }
-    }
-
-    void testEmptyDefault(int x) {
-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
-            case 1:
-                System.out.println("one");
-                break;
-            default:
-                break;
-        }
-    }
 }
