+++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests24/DiffMutationTests24.diff
@@ -1,0 +1,121 @@
+--- src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests24/InputMutationTests24.java
++++ src/test/resources/org/checkstyle/autofix/recipe/useenhancedswitch/mutationtests24/OutputMutationTests24.java
+@@ -7,32 +7,26 @@
+ */
+ package org.checkstyle.autofix.recipe.useenhancedswitch.mutationtests24;
+ 
+-public class InputMutationTests24 {
++public class OutputMutationTests24 {
+     void testContinueControlFlow(int x) {
+         for (int i = 0; i < 10; i++) {
+             int y;
+-            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-                case 1:
++            switch (x) {
++                case 1 -> {
+                     if (i > 5) continue;
+-                    y = 1;
+-                    break;
+-                default:
+-                    y = 0;
+-                    break;
++                    y = 1;}
++                default -> y = 0;
+             }
+         }
+     }
+ 
+     void testReturnControlFlowInAssignment(int x) {
+         int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
++        switch (x) {
++            case 1 -> {
+                 if (x > 100) return;
+-                y = 1;
+-                break;
+-            default:
+-                y = 0;
+-                break;
++                y = 1;}
++            default -> y = 0;
+         }
+     }
+ 
+@@ -40,59 +34,43 @@
+         outer:
+         for (int i = 0; i < 5; i++) {
+             int y;
+-            switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-                case 1:
++            switch (x) {
++                case 1 -> {
+                     if (x > 100) break outer;
+-                    y = 1;
+-                    break;
+-                default:
+-                    y = 0;
+-                    break;
++                    y = 1;}
++                default -> y = 0;
+             }
+         }
+     }
+ 
+     void testBlockIndent(int x) {
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1: {
++        switch (x) {
++            case 1 -> {
+                     int y = 1;
+                     System.out.println(y);
+-                    break;
+                 }
+-            default: {
++            default -> {
+                     int y = 0;
+                     System.out.println(y);
+-                    break;
+                 }
+         }
+     }
+ 
+     void testAssignmentBlockYield(int x) {
+-        int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1: {
++        int y = switch (x) {
++            case 1 -> {
+                 int z = x + 1;
+-                y = z;
+-                break;
++                yield z;
+             }
+-            default:
+-                y = 0;
+-                break;
+-        }
++            default -> 0;
++        };
+     }
+ 
+     void testSingleAssignment(int x) {
+-        int y;
+-        switch (x) { // violation 'Switch can be replaced with enhanced switch.'
+-            case 1:
+-                y = 100;
+-                break;
+-            case 2:
+-                y = 200;
+-                break;
+-            default:
+-                y = 0;
+-                break;
+-        }
++        int y = switch (x) {
++            case 1 -> 100;
++            case 2 -> 200;
++            default -> 0;
++        };
+     }
+ }
